requirements.txt
asgiref==3.6.0
Django==4.2
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
